requirements.txt
google-generativeai==0.3.2
Pillow==10.1.0
python-dotenv==1.0.0
streamlit==1.28.1
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0
